backport of commit 808cdfc2d28afcb2c5285be1355cf31fd7bdc53c
